+++ cognite/neat/rules/_importer/_owl2rules/_owl2properties.py
